Final Commit For Final Project
I should know better than to put 'Final' in any commit by now. A major challenge in this project was to commit the scope of the analysis and application from becoming unmanageable in a tight time-frame. I'll probably move a lot of this work over to a public repository so I can continue to develop it and share my findings with others.